requirements.txt
django
djangorestframework
python-dotenv
requests
pymysql